Add command line parser to project with unit test suite.
At the moment, the command line parser takes in four arguments:
(1) The DNS server ip that will be used for the domain name resolution. This is a required argument.
(2) The domain name that the user wishes to resolve. This is a required argument.
(3) An optional "verbose" argument that, if specified, displays additional tracing/resolution information.
(4) An optional "ipv6" argument that, if specified, resolves the domain name to its appropriate ipv6 address, instead of an ipv4 address.