Move all test data into a test utility class.
This is done for multiple reasons.
(1) Separate code from data (although, to be fair, this data should probably be in some JSON or other data file, rather than in a class).
(2) Enable the use of same test data over many different test suites.
(3) And, presuming the data is used in many different test suites, allow for a single locus of control of the data, if the test data needs to be changed in the future.