Add a random number generator as an instance variable of a ResolverCore object.
I need to expose this implementation detail to allow me to comprehensively test ResolverCore.

Essentially, a random number is generated and set as the query id of the dns messages sent to the dns name servers during the name resolution process. When waiting for a response from the dns name server ResolverCore will check the query id of any dns messages received through socket.recv. To allow for testing, I need to be able to mock the random number generator such that I can dictate exactly which random number is generated to match the query ids of the dns messages I have mocked the socket to return during the name resolution process. Another reminder to always follow TDD, I would have caught this early if I had developed ResolverCore with TDD.